Add Responsive web design to hash rate chart
