+++ d3module/hashrate/index.html
@@ -50,7 +50,7 @@
 
 </head>
 <body>
-<svg id="hashrate" width="900" height="500"></svg>
+<svg id="hashrate"></svg>
 </body>
 
 <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/d3/3.5.10/d3.js"></script>
